+++ internal/middleware/correlation.go
@@ -2,24 +2,11 @@ package middleware
 
 import (
 	"context"
-	"log"
 	"net/http"
 
 	employees_api "github.com/dilyara4949/employees-api"
 	"github.com/google/uuid"
 )
-
-//const CorrelationIDHeader = "X-Correlation-ID"
-
-func Logger() Adapter {
-	return func(h http.Handler) http.HandlerFunc {
-		return func(w http.ResponseWriter, r *http.Request) {
-			id := r.Context().Value(employees_api.CorrelationID)
-			log.Printf("%s %s, correlationID=%v", r.Method, r.URL, id)
-			h.ServeHTTP(w, r)
-		}
-	}
-}
 
 func СorrelationIDMiddleware() Adapter {
 	return func(h http.Handler) http.HandlerFunc {
